Obtain data from the CDC PLACES APIs.

Use this function to access CDC PLACES API data. Measures are sourced from the Behavioral Risk Factor Surveillance System and the American Community Survey ACS.

Usage

get_places(
  geography = "county",
  state = NULL,
  measure = NULL,
  county = NULL,
  release = "2023",
  geometry = FALSE
)

Arguments

geography

The level of desired geography. Currently supports 'county', 'census', and 'zcta'.

state

Specify the state of the desired data using the two letter abbreviation. Supports multiple states if desired.

measure

Specify the measures of the data pull. Supports multiple states if desired. For a full list of available measures, see the function 'get_dictionary'.

county

Specify the county of the desired data using the full name of the county, with a capital letter.

release

Specify the year of release for the PLACES data set. Currently supports years 2020-2023.

geometry

if FALSE (the default), return a regular data frame of PLACES data. If TRUE, uses the tigris package to return an sf data frame with simple feature geometry in the 'geometry' column.

Value

A tibble that contains observations for each measure (age-adjusted and unadjusted prevalence for counties) and geographic level.

Examples

get_places(geography = "county", state = "MI", measure = "SLEEP", release = "2023")
#> # A tibble: 166 × 21
#>    year  stateabbr statedesc locationname datasource category            measure
#>    <chr> <chr>     <chr>     <chr>        <chr>      <chr>               <chr>  
#>  1 2020  MI        Michigan  Cheboygan    BRFSS      Health Risk Behavi… Sleepi…
#>  2 2020  MI        Michigan  Presque Isle BRFSS      Health Risk Behavi… Sleepi…
#>  3 2020  MI        Michigan  Menominee    BRFSS      Health Risk Behavi… Sleepi…
#>  4 2020  MI        Michigan  Allegan      BRFSS      Health Risk Behavi… Sleepi…
#>  5 2020  MI        Michigan  Sanilac      BRFSS      Health Risk Behavi… Sleepi…
#>  6 2020  MI        Michigan  Muskegon     BRFSS      Health Risk Behavi… Sleepi…
#>  7 2020  MI        Michigan  Otsego       BRFSS      Health Risk Behavi… Sleepi…
#>  8 2020  MI        Michigan  Washtenaw    BRFSS      Health Risk Behavi… Sleepi…
#>  9 2020  MI        Michigan  Antrim       BRFSS      Health Risk Behavi… Sleepi…
#> 10 2020  MI        Michigan  Hillsdale    BRFSS      Health Risk Behavi… Sleepi…
#> # ℹ 156 more rows
#> # ℹ 14 more variables: data_value_unit <chr>, data_value_type <chr>,
#> #   data_value <dbl>, low_confidence_limit <dbl>, high_confidence_limit <dbl>,
#> #   totalpopulation <chr>, locationid <chr>, categoryid <chr>, measureid <chr>,
#> #   datavaluetypeid <chr>, short_question_text <chr>, type <chr>, lon <dbl>,
#> #   lat <dbl>
get_places(geography = "county", state = c("MI", "OH"),
measure = c("SLEEP", "ACCESS2"), release = "2023")
#> # A tibble: 684 × 21
#>    year  stateabbr statedesc locationname datasource category            measure
#>    <chr> <chr>     <chr>     <chr>        <chr>      <chr>               <chr>  
#>  1 2020  MI        Michigan  Cheboygan    BRFSS      Health Risk Behavi… Sleepi…
#>  2 2020  MI        Michigan  Presque Isle BRFSS      Health Risk Behavi… Sleepi…
#>  3 2020  MI        Michigan  Menominee    BRFSS      Health Risk Behavi… Sleepi…
#>  4 2020  MI        Michigan  Allegan      BRFSS      Health Risk Behavi… Sleepi…
#>  5 2020  MI        Michigan  Sanilac      BRFSS      Health Risk Behavi… Sleepi…
#>  6 2020  MI        Michigan  Muskegon     BRFSS      Health Risk Behavi… Sleepi…
#>  7 2020  MI        Michigan  Otsego       BRFSS      Health Risk Behavi… Sleepi…
#>  8 2020  MI        Michigan  Washtenaw    BRFSS      Health Risk Behavi… Sleepi…
#>  9 2020  MI        Michigan  Antrim       BRFSS      Health Risk Behavi… Sleepi…
#> 10 2020  MI        Michigan  Hillsdale    BRFSS      Health Risk Behavi… Sleepi…
#> # ℹ 674 more rows
#> # ℹ 14 more variables: data_value_unit <chr>, data_value_type <chr>,
#> #   data_value <dbl>, low_confidence_limit <dbl>, high_confidence_limit <dbl>,
#> #   totalpopulation <chr>, locationid <chr>, categoryid <chr>, measureid <chr>,
#> #   datavaluetypeid <chr>, short_question_text <chr>, type <chr>, lon <dbl>,
#> #   lat <dbl>